api route protection

+++ app/api/items/[id]/route.ts
@@ -1,5 +1,7 @@
 import prisma from "@/app/utils/prismadb";
 import { NextResponse } from "next/server";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "@/app/utils/authOptions";
 
 export async function GET(
   req: Request,
@@ -19,6 +21,10 @@ export async function PUT(
   req: Request,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
+  }
   const { description, imageUrl, price, publicId, selectedCategory } =
     await req.json();
   const id = params.id;
@@ -45,6 +51,10 @@ export async function DELETE(
   req: Request,
   { params }: { params: { id: string } }
 ) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: "Not authenticated" }, { status: 401 });
+  }
   try {
     const id = params.id;
     const deletedItem = await prisma.item.delete({ where: { id } });
